fix : fix type compatible error

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: rjeong <rjeong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/06 20:12:45 by rjeong            #+#    #+#             */
-/*   Updated: 2022/07/07 19:45:45 by rjeong           ###   ########.fr       */
+/*   Updated: 2022/07/09 21:45:12 by rjeong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
